rag_agent/requirements.txt
google-adk
google-cloud-aiplatform[adk,agent_engines]
google-cloud-storage
google-genai
gitpython
streamlit
requests
sseclient-py
python-dotenv
